Better format binary expressions
